+++ app/src/main/java/com/example/sh/androidregisterandlogin/TotalHome/Frags/PhotoFragment.java
@@ -1,6 +1,5 @@
-package com.example.sh.androidregisterandlogin.ToTalHome.Frags;
+package com.example.sh.androidregisterandlogin.TotalHome.Frags;
 
-import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
@@ -15,12 +14,12 @@ import com.example.sh.androidregisterandlogin.R;
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link Test1Fragment.OnFragmentInteractionListener} interface
+ * {@link PhotoFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link Test1Fragment#newInstance} factory method to
+ * Use the {@link PhotoFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Test1Fragment extends Fragment {
+public class PhotoFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -31,20 +30,25 @@ public class Test1Fragment extends Fragment {
     private String mParam2;
 
     private OnFragmentInteractionListener mListener;
-    public static Test1Fragment newInstance() {
-        return new Test1Fragment();
+
+    public static PhotoFragment newInstance() {
+        return new PhotoFragment();
     }
+    public PhotoFragment() {
+        // Required empty public constructor
+    }
+
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Test1Fragment.
+     * @return A new instance of fragment PhotoFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static Test1Fragment newInstance(String param1, String param2) {
-        Test1Fragment fragment = new Test1Fragment();
+    public static PhotoFragment newInstance(String param1, String param2) {
+        PhotoFragment fragment = new PhotoFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -65,7 +69,7 @@ public class Test1Fragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_test1, container, false);
+        return inflater.inflate(R.layout.fragment_photo, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
@@ -74,7 +78,7 @@ public class Test1Fragment extends Fragment {
             mListener.onFragmentInteraction(uri);
         }
     }
-//
+
 //    @Override
 //    public void onAttach(Context context) {
 //        super.onAttach(context);
