+++ app/src/main/java/com/example/sh/androidregisterandlogin/TotalHome/Frags/Test1Fragment.java
@@ -1,6 +1,5 @@
-package com.example.sh.androidregisterandlogin.ToTalHome.Frags;
+package com.example.sh.androidregisterandlogin.TotalHome.Frags;
 
-import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
@@ -15,12 +14,12 @@ import com.example.sh.androidregisterandlogin.R;
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link Test4Fragment.OnFragmentInteractionListener} interface
+ * {@link Test1Fragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link Test4Fragment#newInstance} factory method to
+ * Use the {@link Test1Fragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Test4Fragment extends Fragment {
+public class Test1Fragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -31,13 +30,8 @@ public class Test4Fragment extends Fragment {
     private String mParam2;
 
     private OnFragmentInteractionListener mListener;
-
-    public Test4Fragment() {
-        // Required empty public constructor
-    }
-
-    public static Test4Fragment newInstance() {
-        return new Test4Fragment();
+    public static Test1Fragment newInstance() {
+        return new Test1Fragment();
     }
     /**
      * Use this factory method to create a new instance of
@@ -45,11 +39,11 @@ public class Test4Fragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Test4Fragment.
+     * @return A new instance of fragment Test1Fragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static Test4Fragment newInstance(String param1, String param2) {
-        Test4Fragment fragment = new Test4Fragment();
+    public static Test1Fragment newInstance(String param1, String param2) {
+        Test1Fragment fragment = new Test1Fragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -70,7 +64,7 @@ public class Test4Fragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_test4, container, false);
+        return inflater.inflate(R.layout.fragment_test1, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
