+++ app/src/main/java/com/example/sh/androidregisterandlogin/TotalHome/Frags/Test4Fragment.java
@@ -1,6 +1,5 @@
-package com.example.sh.androidregisterandlogin.ToTalHome.Frags;
+package com.example.sh.androidregisterandlogin.TotalHome.Frags;
 
-import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 
@@ -15,12 +14,12 @@ import com.example.sh.androidregisterandlogin.R;
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link Test2Fragment.OnFragmentInteractionListener} interface
+ * {@link Test4Fragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link Test2Fragment#newInstance} factory method to
+ * Use the {@link Test4Fragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Test2Fragment extends Fragment {
+public class Test4Fragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -32,24 +31,24 @@ public class Test2Fragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public static Test2Fragment newInstance() {
-        return new Test2Fragment();
-    }
-    public Test2Fragment() {
+    public Test4Fragment() {
         // Required empty public constructor
     }
 
+    public static Test4Fragment newInstance() {
+        return new Test4Fragment();
+    }
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Test2Fragment.
+     * @return A new instance of fragment Test4Fragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static Test2Fragment newInstance(String param1, String param2) {
-        Test2Fragment fragment = new Test2Fragment();
+    public static Test4Fragment newInstance(String param1, String param2) {
+        Test4Fragment fragment = new Test4Fragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -70,7 +69,7 @@ public class Test2Fragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_test2, container, false);
+        return inflater.inflate(R.layout.fragment_test4, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
@@ -79,7 +78,7 @@ public class Test2Fragment extends Fragment {
             mListener.onFragmentInteraction(uri);
         }
     }
-
+//
 //    @Override
 //    public void onAttach(Context context) {
 //        super.onAttach(context);
